feat(lyrics): BitChord's bottom bar for every style, and no more hiding controls
The lyrics screen's controls hid themselves after five seconds and had to be
poked back into existence, behind two preferences. The BitChord style instead
kept a persistent bar under its lyrics — scrubber, elapsed and remaining, the
quality badge, and the lyrics provider beside the way out. That bar is now what
every lyrics style shows, and nothing hides.

What the bar gains over the old controls:

- The quality badge between the timestamps ("Lossless", "Hi-Res Lossless",
  "Hi-Quality", or "Upgrading Quality" while a better stream is still
  resolving). Centred on the row rather than placed in the gap between the two
  times, because that gap changes width by a digit every time a minute rolls
  over and a badge that shifts with the clock reads as a glitch.
- The provider chip — "Lyrics from X" — which is also the way into the lyric
  actions, with the same ⋯ beside it, and a ✕ to leave.

The transport row stays and is simply always visible; the answer to controls
that hide badly is not to remove them.

LosslessOrStats moves out of the BitChord package into ui.player as FormatBadge.
It is shared UI now, and a shared lyrics screen reaching into one style's package
for it would be the wrong way round.

Removed with the behaviour: ShowLyricsPlayerControlsKey and
AutoHideLyricsPlayerControlsKey, their two switches in Lyrics settings, their two
entries in settings search, the two toggles inside LyricsMenu and the
showControlsToggles parameter that existed to hide them from the styles that
never wanted them, the five-second timers in both LyricsScreen and
AppleMusicPlayer, and the poke-to-reveal gesture plumbing that reached from the
root container down into every transport button. The orphaned strings go too, in
all four locales.

pokePlayerControlsVisibility survives in AppleMusicPlayer as a no-op rather than
being threaded out of a dozen components that report a tap — the call sites are
harmless and removing them is a larger diff than leaving them.

Ponytail (applied by hand — the plugin is not loadable in this session): the two
flags become `true` and the dead branches collapse, rather than a "controls
always visible" preference being added to replace the two being removed.
