+++ app/src/main/kotlin/moe/rukamori/archivetune/ui/player/bitchord/BitChordPlayer.kt
@@ -186,6 +186,7 @@ import moe.rukamori.archivetune.constants.LyricsModeKey
 import moe.rukamori.archivetune.ui.component.LyricsEnhanced
 import moe.rukamori.archivetune.ui.component.LyricsV2
 import moe.rukamori.archivetune.utils.rememberEnumPreference
+import moe.rukamori.archivetune.ui.player.LosslessOrStats
 import androidx.compose.material.icons.rounded.Close
 import androidx.compose.material.icons.rounded.Headphones
 import androidx.compose.runtime.MutableFloatState
@@ -1541,11 +1542,7 @@ fun BitChordPlayerContent(
                                         mediaMetadataProvider = { mediaMetadata },
                                         lyricsSyncOffset = lyricsSyncOffset,
                                         onLyricsSyncOffsetChange = { lyricsSyncOffset = it },
-                                        showPlayerControlsState = null,
-                                        onShowPlayerControlsChange = null,
-                                        onAutoHidePlayerControlsChange = {},
                                         onDismiss = menuState::dismiss,
-                                        showControlsToggles = false,
                                     )
                                 }
                             },
@@ -1947,157 +1944,6 @@ private fun formatTime(ms: Long): String {
     val minutes = TimeUnit.MILLISECONDS.toMinutes(ms)
     val seconds = TimeUnit.MILLISECONDS.toSeconds(ms) % 60
     return "%d:%02d".format(Locale.ROOT, minutes, seconds)
-}
-
-// ── The codec / quality badge (adapted to ArchiveTune's FormatEntity) ─────────
-
-/**
- * The gap between the two timestamps under the seek bar: just the "Lossless"
- * badge when one applies, and nothing otherwise. The measured stats line lives
- * inside the sleeve instead — the badge is a claim, the sleeve is where the
- * evidence is.
- */
-@Composable
-private fun LosslessOrStats(
-    isLoading: Boolean,
-    format: FormatEntity?,
-    modifier: Modifier = Modifier,
-) {
-    val lossless = format?.isLossless() == true
-    val hiRes = lossless && (format?.sampleRate ?: 0) >= 88_200
-    val hiQuality = !lossless && (format?.bitrate ?: 0) >= 250_000
-    when {
-        // Still resolving — nothing measured yet to confirm with, so this is a
-        // statement of intent, not a result.
-        format == null || (isLoading && !lossless) -> LosslessLabel(
-            text = "Upgrading Quality",
-            animated = false,
-            modifier = modifier,
-        )
-        lossless -> LosslessLabel(
-            // Same line Tidal, Qobuz and Apple Music draw it at.
-            text = if (hiRes) "Hi-Res Lossless" else "Lossless",
-            // Shimmer is reserved for the thing that was asked for and
-            // confirmed. It is what makes the badge read as an achievement
-            // rather than a label, which only one of these two is.
-            animated = true,
-            modifier = modifier,
-        )
-        // Lossy, but the good end of lossy.
-        hiQuality -> LosslessLabel(
-            text = "Hi-Quality",
-            animated = false,
-            modifier = modifier,
-        )
-        else -> {}
-    }
-}
-
-/** Whether the stream is a lossless codec. */
-private fun FormatEntity.isLossless(): Boolean =
-    mimeType.endsWith("flac") || mimeType.endsWith("alac")
-
-/**
- * "FLAC · 320 kbps · 48.0 kHz" — whichever of those the format actually
- * reports. A figure it hasn't is dropped rather than filled in, so a short
- * line means little was known, never that something was invented.
- */
-internal fun FormatEntity.describe(): String {
-    val parts = buildList {
-        codecLabel(mimeType)?.let(::add)
-        if (!isLossless()) add("${bitrate / 1000} kbps")
-        sampleRate?.let { add("%.1f kHz".format(Locale.ROOT, it / 1000f)) }
-    }
-    return parts.joinToString(" · ").takeIf { it.isNotEmpty() } ?: ""
-}
-
-/** The codec under its usual name rather than its MIME type. */
-internal fun codecLabel(mimeType: String?): String? = when {
-    mimeType == null -> null
-    mimeType.endsWith("opus") -> "Opus"
-    mimeType.endsWith("mp4a-latm") -> "AAC"
-    mimeType.endsWith("vorbis") -> "Vorbis"
-    mimeType.endsWith("mpeg") -> "MP3"
-    mimeType.endsWith("flac") -> "FLAC"
-    mimeType.endsWith("alac") -> "ALAC"
-    else -> mimeType.substringAfter('/').uppercase(Locale.ROOT)
-}
-
-/** A headphone glyph ahead of the quality tag — "Upgrading Quality", "Hi-Quality", "Lossless". */
-@Composable
-private fun LosslessLabel(text: String, animated: Boolean, modifier: Modifier = Modifier) {
-    Row(
-        modifier = modifier,
-        horizontalArrangement = androidx.compose.foundation.layout.Arrangement.Center,
-        verticalAlignment = Alignment.CenterVertically,
-    ) {
-        Icon(
-            imageVector = Icons.Rounded.Headphones,
-            contentDescription = null,
-            tint = Color.White.copy(alpha = if (animated) 0.7f else 0.45f),
-            modifier = Modifier.size(13.dp),
-        )
-        Spacer(Modifier.width(4.dp))
-        if (animated) {
-            ShimmerText(text = text)
-        } else {
-            Text(
-                text = text,
-                style = MaterialTheme.typography.labelMedium.copy(
-                    fontSize = (MaterialTheme.typography.labelMedium.fontSize.value + 1).sp,
-                ),
-                color = Color.White.copy(alpha = 0.45f),
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-    }
-}
-
-/**
- * "Lossless", with a highlight band sweeping left to right across it every
- * three seconds — confirmed, not just claimed, so it's worth the shine.
- *
- * The band's width is measured off the text itself via [onSizeChanged]
- * rather than assumed, so the sweep always clears the word fully at both
- * ends instead of being sized for whatever length happened to be typical.
- */
-@Composable
-private fun ShimmerText(text: String) {
-    var widthPx by remember { mutableIntStateOf(0) }
-    val transition = androidx.compose.animation.core.rememberInfiniteTransition(label = "lossless-shimmer")
-    val progress by transition.animateFloat(
-        initialValue = 0f,
-        targetValue = 1f,
-        animationSpec = androidx.compose.animation.core.infiniteRepeatable(
-            animation = tween(durationMillis = 3_000, easing = androidx.compose.animation.core.LinearEasing),
-            repeatMode = androidx.compose.animation.core.RepeatMode.Restart,
-        ),
-        label = "lossless-shimmer-progress",
-    )
-    val baseColor = Color.White.copy(alpha = 0.55f)
-    val brush = if (widthPx <= 0) {
-        Brush.linearGradient(listOf(baseColor, baseColor))
-    } else {
-        val band = widthPx * 0.6f
-        val center = -band + progress * (widthPx + 2 * band)
-        Brush.linearGradient(
-            colorStops = arrayOf(0f to baseColor, 0.5f to Color.White, 1f to baseColor),
-            start = Offset(center - band, 0f),
-            end = Offset(center + band, 0f),
-        )
-    }
-    Text(
-        text = text,
-        style = MaterialTheme.typography.labelMedium.copy(
-            brush = brush,
-            fontWeight = FontWeight.SemiBold,
-            fontSize = (MaterialTheme.typography.labelMedium.fontSize.value + 1).sp,
-        ),
-        maxLines = 1,
-        overflow = TextOverflow.Ellipsis,
-        modifier = Modifier.onSizeChanged { widthPx = it.width },
-    )
 }
 
 // ── OverlayBack (verbatim) ────────────────────────────────────────────────────
