+++ app/src/main/kotlin/moe/rukamori/archivetune/ui/screens/settings/LyricsSettings.kt
@@ -67,7 +67,6 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
 import moe.rukamori.archivetune.LocalPlayerAwareWindowInsets
 import moe.rukamori.archivetune.R
-import moe.rukamori.archivetune.constants.AutoHideLyricsPlayerControlsKey
 import moe.rukamori.archivetune.constants.EnableBetterLyricsKey
 import moe.rukamori.archivetune.constants.EnableBetterLyricsPortatoKey
 import moe.rukamori.archivetune.constants.EnableKugouKey
@@ -85,7 +84,6 @@ import moe.rukamori.archivetune.constants.LyricsRomanizeJapaneseKey
 import moe.rukamori.archivetune.constants.LyricsRomanizeKoreanKey
 import moe.rukamori.archivetune.constants.LyricsRomanizeOtherLanguagesKey
 import moe.rukamori.archivetune.constants.LyricsScrollKey
-import moe.rukamori.archivetune.constants.ShowLyricsPlayerControlsKey
 import moe.rukamori.archivetune.constants.LyricsTextSizeKey
 import moe.rukamori.archivetune.constants.PreferredLyricsProvider
 import moe.rukamori.archivetune.constants.QueueLyricsPreloadCountKey
@@ -161,10 +159,6 @@ fun LyricsSettings(
             deserializeLyricsProviderOrder(providerOrderStr)
         }
     val (lyricsLineBlur, onLyricsLineBlurChange) = rememberPreference(LyricsLineBlurKey, defaultValue = false)
-    // The "Show player controls" and "Auto-hide controls" toggles were removed from the UI
-    // by user request — the bottom player controls now ALWAYS auto-hide after 5 s in the
-    // Apple Music style lyrics view, with no opt-out. The backing preference keys are kept
-    // (and forced to `true` in the player code) so existing installs don't break on upgrade.
     val (lyricsRomanizeJapanese, onLyricsRomanizeJapaneseChange) = rememberPreference(LyricsRomanizeJapaneseKey, defaultValue = false)
     val (lyricsRomanizeKorean, onLyricsRomanizeKoreanChange) = rememberPreference(LyricsRomanizeKoreanKey, defaultValue = true)
     val (lyricsRomanizeChinese, onLyricsRomanizeChineseChange) = rememberPreference(LyricsRomanizeChineseKey, defaultValue = true)
@@ -421,13 +415,12 @@ fun LyricsSettings(
                 )
             }
 
-            // ── Removed by user request ──────────────────────────────────────────
+            // ── Removed (2026-09-04 port of 6701ec36b) ────────────────────────────
             // "Show player controls" and "Auto-hide controls" toggles used to live here.
-            // Both are now hardcoded on — the Apple Music lyrics view always shows the
-            // bottom controls and always auto-hides them after 5 s (tap anywhere to bring
-            // them back). The preferences (ShowLyricsPlayerControlsKey,
-            // AutoHideLyricsPlayerControlsKey) are kept in PreferenceKeys.kt so legacy
-            // installs don't see a DataStore corruption error on upgrade.
+            // The lyrics controls no longer hide at all: the bottom bar carries the
+            // scrubber, the quality badge and the lyrics provider, so there is nothing
+            // left to toggle. The preference keys were removed from PreferenceKeys.kt
+            // with the rest of the plumbing.
             // ─────────────────────────────────────────────────────────────────────
 
             item {
