+++ app/src/main/res/values-pt-rBR/archivetune_strings.xml
@@ -598,7 +598,6 @@
     <string name="music_recognition_notification_recording_failed_text">O aplicativo em execução pode estar bloqueando a captura de áudio, ou a reprodução pode ter sido interrompida.</string>
     <string name="music_recognition_notification_signature_failed_text">O áudio capturado não continha música suficiente para ser identificado.</string>
     <string name="music_recognition_notification_recognition_failed_text">O serviço de reconhecimento não conseguiu identificar este áudio. Tente novamente daqui a pouco.</string>
-    <string name="show_lyrics_player_controls">Mostrar controles do player</string>
     <string name="block_artist">Bloquear artista</string>
     <string name="unblock_artist">Desbloquear artista</string>
     <string name="stats_top_songs">Top músicas</string>
